% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_hack.R
\name{.clusterExport_function}
\alias{.clusterExport_function}
\title{Internal function to enclose environment}
\usage{
.clusterExport_function(cl, FUN)
}
\arguments{
\item{cl}{(vector (atomic or list)) the vector to loop on.}

\item{FUN}{the function to be applied to each element of \code{X} in parallel to.}
}
\description{
A script called by mclapply_hack to enclose environment.
}
