% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_hack.R
\name{mclapply_hack}
\alias{mclapply_hack}
\title{Windows version of parallel apply function to use}
\usage{
mclapply_hack(...)
}
\arguments{
\item{...}{Any inputs works for \code{makeCluster}.}
}
\description{
A script to implement a hackish version of parallel:mclapply()
on Windows machines.
}
