% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Re-class the target map}
\usage{
reclass(input, output, nthread = detectCores())
}
\arguments{
\item{input}{(SpatRaster) an input terra raster to resample}

\item{output}{(SpatRaster) the output grid to resample to}

\item{nthread}{(integer) the number of thread to use for parallel.
Default to \code{\link{detectCores}}.}
}
\value{
(SpatRaster) the target terra raster
}
\description{
re-class the base map to the target map
}
\examples{
library(APUpscale)
library(terra)
nlcd <- rast(system.file('extdata/nlcd_dukes.tif', package = "APUpscale"))
output <- rast(ext(nlcd), crs = crs(nlcd), resolution = 1000)
lc_coarse <- reclass(nlcd, output, nthread = 2)
coltab(lc_coarse) <- coltab(nlcd)
plot(lc_coarse)
}
