% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale.R
\name{upscale}
\alias{upscale}
\title{upscale}
\usage{
upscale(input, cellsize, no_data = NA, nthread = detectCores())
}
\arguments{
\item{input}{(SpatRaster) a terra raster with fine resolution}

\item{cellsize}{(numeric) cell resolution in units of input CRS of
output target grid.}

\item{no_data}{(integer) a value to be treated as NO_DATA. Default to NA.}

\item{nthread}{(integer) the number of thread to use for parallel.
Default to \code{parallel::detectCores()}.}
}
\value{
(SpatRaster) a resampled terra raster
}
\description{
upscale an input raster to an output grid
using an area preserving method
}
