% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_hack.R
\name{mclapply_hack}
\alias{mclapply_hack}
\title{Windows version of parallel apply function to use}
\usage{
mclapply_hack(X, FUN, ...)
}
\arguments{
\item{X}{(vector (atomic or list)) the vector to loop on.}

\item{FUN}{the function to be applied to each element of \code{X} in parallel to.}

\item{...}{Any inputs works for \code{makeCluster}.}
}
\description{
A script to implement a hackish version of parallel:mclapply()
on Windows machines.
}
\examples{
library(APUpscale)
library(parallel)
mclapply <- switch( Sys.info()[['sysname']],
                    Windows = {mclapply_hack},
                    Linux   = {mclapply},
                    Darwin  = {mclapply})
}
