% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Re-class the target map}
\usage{
reclass(input, output, nthread = detectCores())
}
\arguments{
\item{input}{(SpatRaster) an input terra raster to resample}

\item{output}{(SpatRaster) the output grid to resample to}

\item{nthread}{(integer) the number of thread to use for parallel.
Default to \code{parallel::detectCores()}.}
}
\value{
(SpatRaster) the target terra raster
}
\description{
re-class the base map to the target map
}
