% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Re-class the target map}
\usage{
reclass(input, output, nthread = NULL, verbose = FALSE)
}
\arguments{
\item{input}{(SpatRaster) an input terra raster to resample}

\item{output}{(SpatRaster) the output grid to resample to}

\item{nthread}{(integer) the number of thread to use for parallel.
Default to \code{NULL}. It is recommended to use small ones to
take full advantage of \code{terra}, for instance 2, 4.}

\item{verbose}{(logical) print out info for debugging.}
}
\value{
(matrix) the categorical matrix of target terra raster
}
\description{
re-class the base map to the target map
}
\examples{
library(APUpscale)
library(terra)
nlcd <- rast(system.file('extdata/nlcd_dukes.tif', package = "APUpscale"))
output <- rast(ext(nlcd), crs = crs(nlcd), resolution = 1000, vals = NA)
target_vals <- reclass(nlcd, output, nthread = 2)
values(output) <- target_vals
coltab(output) <- coltab(nlcd)
plot(output)
}
