% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale.R
\name{upscale}
\alias{upscale}
\title{upscale}
\usage{
upscale(input, cellsize, no_data = NA, nthread = NULL, verbose = FALSE)
}
\arguments{
\item{input}{(SpatRaster) a terra raster with fine resolution}

\item{cellsize}{(numeric) cell resolution in units of input CRS of
output target grid. The value should be always larger than input resolution.}

\item{no_data}{(integer) a value to be treated as NO_DATA. Default to NA.}

\item{nthread}{(integer) the number of thread to use for parallel. If \code{NULL},
the all available cores in the machine will be used. Default to \code{NULL}.}

\item{verbose}{(logical) option to print out info for debugging.}
}
\value{
(SpatRaster) a resampled terra raster
}
\description{
upscale an input raster to an output grid
using an area preserving method
}
\examples{
library(APUpscale)
library(terra)
nlcd <- rast(system.file('extdata/nlcd_dukes.tif', package = "APUpscale"))
lc_coarse <- upscale(nlcd, cellsize = 1000, nthread = 2)
coltab(lc_coarse) <- coltab(nlcd)
plot(lc_coarse)
}
